Create README.md
Right now I am just practicing with the case 1 numbers. My idea is to take the top player (I put them in order for this example) and figure out the difference between them and the next guy down, keep doing that for however many players the team needs. Add those up. Those get put into an array. Then you go to the 2nd best and so on until you go through all possible combinations. The loop I am working on is to create that array. Afterwards I'll have it indicate the smallest number in the array which is the answer.